pretty much a rewrite.
we no longer claim to have anything to do with vt220.  it's ecma-48
now, or really an old ansi standard perhaps.  we only support as
few escape sequences as we can get away with (and that's allowed
by ecma-48).

we now do text selection with copy & paste.  we can do resize by
writing a ctl message to ssh's /dev/termctl.  it's a bit slow, but
very usable.  this is committed in vim in novt in hosted inferno.